+++ package.json
@@ -33,10 +33,10 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "../react-netflix-clone/node_modules/.bin/react-scripts start",
-    "build": "../react-netflix-clone/node_modules/.bin/react-scripts build",
-    "test": "../react-netflix-clone/node_modules/.bin/react-scripts test",
-    "eject": "../react-netflix-clone/node_modules/.bin/react-scripts eject"
+    "start": "../junggkim/node_modules/.bin/react-scripts start",
+    "build": "../junggkim/node_modules/.bin/react-scripts build",
+    "test": "../junggkim/node_modules/.bin/react-scripts test",
+    "eject": "../junggkim/node_modules/.bin/react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
