+++ src/components/Nav.js
@@ -1,4 +1,3 @@
-
 import React, {useContext, useEffect, useState} from "react";
 import "./Nav.css";
 import LoginDropdown from "./recommendFriend/LoginDropdown";
@@ -8,19 +7,24 @@ import logo from "./search/assets/waffle.png";
 import {RecommendationContext} from "./recommendMain/RecommendationContext";
 import "./recommendMain/styles/Header.css";
 import profileLogo from "./recommendFriend/img/ProfileLogo.png";
+import InviteModal from "./recommendFriend/InviteModal";
+import {autoRefreshCheck} from "../tokenUtils/TokenUtils";
 
 export default function Nav({
-                              onClickRecommend,
-                              isLoggedIn2,
-                              onLoginClick,
-                              onLogout,
-                              onProfileClick,
-                              showProfileDropdown,
+                                onClickRecommend,
+                                isLoggedIn2,
+                                onLoginClick,
+                                onLogout,
+                                onProfileClick,
+                                showProfileDropdown,
                             }) {
     const [search, setSearch] = useState("");
-  const [show, setShow] = useState(false);
-  const [showLoginDropdown, setShowLoginDropdown] = useState(false);
+    const [show, setShow] = useState(false);
+    const [showLoginDropdown, setShowLoginDropdown] = useState(false);
     const navigate = useNavigate();
+// 자동 친구 추가 기능 state 추가
+    const [invites, setInvites] = useState([]);
+    const [showInviteModal, setShowInviteModal] = useState(false);
 
     const {
         clearRecommendations,
@@ -46,83 +50,151 @@ export default function Nav({
     };
 
 
-  useEffect(() => {
-    const handleScroll = () => {
-      setShow(window.scrollY > 50);
+    useEffect(() => {
+        const handleScroll = () => {
+            setShow(window.scrollY > 50);
+        };
+
+        window.addEventListener("scroll", handleScroll);
+        return () => window.removeEventListener("scroll", handleScroll);
+    }, []);
+
+    // 로그인 후 드롭다운 닫기와 상태 변경을 한 번에 처리
+    const handleLoginSuccess = async () => {
+        onLoginClick(); // RecommendReal 상태 변경
+        setShowLoginDropdown(false); // 드롭다운 닫기
+
+        await tryInviteFriend()
     };
 
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
+    const tryInviteFriend = async () => {
+        // uuid로 시작하는 토큰들 싸그리 모아서 확인한다.
+        const uuidTokens = [];
 
-  // 로그인 후 드롭다운 닫기와 상태 변경을 한 번에 처리
-  const handleLoginSuccess = () => {
-    onLoginClick(); // RecommendReal 상태 변경
-    setShowLoginDropdown(false); // 드롭다운 닫기
-  };
+        for (let i = 0; i < localStorage.length; i++) {
+            const key = localStorage.key(i);
+            if (key.startsWith("uuid")) {
+                uuidTokens.push({
+                    key: key,
+                    value: localStorage.getItem(key),
+                });
+            }
+        }
+        if (uuidTokens.length < 1) {
 
-  return (
-      <nav className={`nav2 ${show && "nav2__black"}`}>
-        <div className="nav__left">
-          <img
-              alt="waffle logo"
-              src={logo}
-              className="nav__logo"
-              onClick={() => navigate('/')}
-          />
-            <Link style={{textDecoration: "none"}} to={"/search"}><h2 className="nav__recommend" >
-                검색
-            </h2></Link>
-            <Link style={{textDecoration: "none"}} to={"/recommend/main"}><h2 className="nav__recommend" >
-            추천
-          </h2></Link>
-            <Link style={{textDecoration: "none"}} to={"/recommend/friend"}><h2 className="nav__recommend" >
-                추천 with Friends
-            </h2></Link>
+            return;
+        }
+        try {
+            const response = await autoRefreshCheck({
+                url: "http://localhost:8080/api/friend/nicknameByUuids",
+                method: "POST",
+                data: {
+                    uuids: uuidTokens.map((item) => item.value),
+                },
+                headers: {
+                    "Content-Type": "application/json",
+                },
+            });
 
-            <Link style={{textDecoration: "none"}} to={"/post"}><h2 className="nav__recommend" >
-                게시판
-            </h2></Link>
-        </div>
+            const enriched = uuidTokens.map((item) => ({
+                key: item.key,
+                uuid: item.value,
+                nickname: response.data[item.value] || "알 수 없음",
+            }));
 
-          <button
-              className={`header-icon-button member-mode-toggle ${isMemberModeActive ? 'active glowing-rainbow' : ''}`}
-              onClick={toggleMemberMode}
-              title={isMemberModeActive ? '회원 모드 활성화됨' : '회원 모드 비활성화됨'}
-          >
-              <i className={`fas ${isMemberModeActive ? 'fa-user-check' : 'fa-user'}`}></i>
-              {isLoggedIn && <span className="member-mode-text">{isMemberModeActive ? '회원 모드' : '일반 모드'}</span>}
-              {!isLoggedIn && <span className="member-mode-text">로그인</span>}
-          </button>
 
-        <div className="nav__right">
-          {!isLoggedIn2 ? (
-              <>
-                <button
-                    className="nav__login"
-                    onClick={() => setShowLoginDropdown(!showLoginDropdown)}
-                >
-                  로그인
-                </button>
-                {showLoginDropdown && (
-                    <LoginDropdown
-                        onClose={() => setShowLoginDropdown(false)}
-                        onLoginSuccess={handleLoginSuccess}
-                    />
-                )}
-              </>
-          ) : (
-              <>
+            setInvites(enriched);
+            setShowInviteModal(true);
+
+        } catch (error) {
+            console.error("초대 닉네임 로드 실패:", error);
+        }
+    };
+
+    function handleRecommend() {
+        const jwt = localStorage.getItem("jwt");
+        if (jwt) {
+            navigate("/recommend/friend");
+        } else {
+            alert("로그인이 필요합니다.")
+        }
+    }
+
+    return (
+        <nav className={`nav2 ${show && "nav2__black"}`}>
+            <div className="nav__left">
                 <img
-                    alt="User profile"
-                    src={profileLogo}
-                    className="nav__avater"
-                    onClick={onProfileClick}
+                    alt="waffle logo"
+                    src={logo}
+                    className="nav__logo"
+                    onClick={() => navigate('/')}
                 />
-                {showProfileDropdown && <ProfileDropdown onLogout={onLogout} />}
-              </>
-          )}
-        </div>
-      </nav>
-  );
+                <Link style={{textDecoration: "none"}} to={"/search"}><h2 className="nav__recommend">
+                    검색
+                </h2></Link>
+                <Link style={{textDecoration: "none"}} to={"/recommend/main"}><h2 className="nav__recommend">
+                    추천
+                </h2></Link>
+                <h2 className="nav__recommend" onClick={handleRecommend}>
+                    추천 with Friends
+                </h2>
+
+                <Link style={{textDecoration: "none"}} to={"/post"}><h2 className="nav__recommend">
+                    게시판
+                </h2></Link>
+            </div>
+
+            {/*자동 친구 추가 모달 창 부분*/}
+            {showInviteModal && (
+                <InviteModal
+                    invites={invites}
+                    onAccept={(acceptedKey) => {
+                        setInvites((prev) => prev.filter((inv) => inv.key !== acceptedKey));
+                    }}
+                    onClose={() => {
+                        setShowInviteModal(false);
+
+                    }}
+                />
+            )}
+            <button
+                className={`header-icon-button member-mode-toggle ${isMemberModeActive ? 'active glowing-rainbow' : ''}`}
+                onClick={toggleMemberMode}
+                title={isMemberModeActive ? '회원 모드 활성화됨' : '회원 모드 비활성화됨'}
+            >
+                <i className={`fas ${isMemberModeActive ? 'fa-user-check' : 'fa-user'}`}></i>
+                {isLoggedIn && <span className="member-mode-text">{isMemberModeActive ? '회원 모드' : '일반 모드'}</span>}
+                {!isLoggedIn && <span className="member-mode-text">로그인</span>}
+            </button>
+
+            <div className="nav__right">
+                {!isLoggedIn2 ? (
+                    <>
+                        <button
+                            className="nav__login"
+                            onClick={() => setShowLoginDropdown(!showLoginDropdown)}
+                        >
+                            로그인
+                        </button>
+                        {showLoginDropdown && (
+                            <LoginDropdown
+                                onClose={() => setShowLoginDropdown(false)}
+                                onLoginSuccess={handleLoginSuccess}
+                            />
+                        )}
+                    </>
+                ) : (
+                    <>
+                        <img
+                            alt="User profile"
+                            src={profileLogo}
+                            className="nav__avater"
+                            onClick={onProfileClick}
+                        />
+                        {showProfileDropdown && <ProfileDropdown onLogout={onLogout}/>}
+                    </>
+                )}
+            </div>
+        </nav>
+    );
 }
