--- src/components/recommendFriend/Nav.css
@@ -1,72 +1,0 @@
-
-.nav2 {
-    position: fixed;
-    top: 0;
-    width: 100%;
-    height: 30px;
-    z-index: 1000;
-    padding: 20px 40px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    transition-timing-function: ease-in;
-    transition: all 0.5s;
-    background-color: transparent;
-}
-
-.nav2__black {
-    background-color: #111;
-}
-
-.nav__left {
-    display: flex;
-    align-items: center;
-}
-
-.nav__logo {
-    width: 200px;
-    object-fit: contain;
-    cursor: pointer;
-    margin-right: 20px;
-}
-
-.nav__recommend {
-    color: #fff;
-    font-size: 20px;
-    cursor: pointer;
-    margin-right: 5px; /* 추천과 프로필 로고 간격 */
-}
-
-.nav__avater {
-    width: 40px;
-    height: 40px;
-    object-fit: cover;
-    cursor: pointer;
-    margin-left: 2px;
-    margin-right: 70px; /* ← 이거 추가해서 살짝 왼쪽으로 이동 */
-    flex-shrink: 0;
-
-}
-
-.nav__right {
-    display: flex;
-    align-items: center;
-    gap: 10px; /* 💡 요소 간 여백은 gap으로 조절 */
-    position: relative;
-    overflow: visible; /* 혹시라도 가려지는 거 방지 */
-    padding-right: 0; /* 추가 */
-}
-
-.nav__login {
-    background: none;
-    border: none;
-    color: #fff;
-    font-size: 18px;
-    cursor: pointer;
-    padding: 10px 20px;
-    margin-right: 40px;
-}
-
-.nav__login:hover {
-    color: #b3b3b3;
-}
