+++ src/tokenUtils/TokenUtils.js
@@ -295,8 +295,7 @@ export async function autoRefreshCheck(config) {
     const token = localStorage.getItem("jwt");
 
     if (!token) {
-        alert("로그인이 필요합니다. 로그인해주세요!");
-        return Promise.reject("No JWT token");
+        return ;
     }
 
     try {
